% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{selfie}
\alias{selfie}
\alias{flexed_biceps}
\alias{backhand_index_pointing_left}
\alias{backhand_index_pointing_right}
\alias{backhand_index_pointing_down}
\alias{backhand_index_pointing_up}
\alias{index_pointing_up}
\alias{middle_finger}
\alias{victory_hand}
\alias{crossed_fingers}
\alias{vulcan_salute}
\alias{sign_of_the_horns}
\alias{call_me_hand}
\alias{hand_with_fingers_splayed}
\alias{raised_hand}
\alias{thumbs_up}
\alias{thumbs_down}
\alias{raised_fist}
\alias{oncoming_fist}
\alias{left_facing_fist}
\alias{right_facing_fist}
\alias{raised_back_of_hand}
\alias{waving_hand}
\alias{love_you_gesture}
\alias{writing_hand}
\alias{clapping_hands}
\alias{open_hands}
\alias{raising_hands}
\alias{palms_up_together}
\alias{folded_hands}
\alias{handshake}
\alias{nail_polish}
\alias{ear}
\alias{nose}
\alias{footprints}
\alias{eyes}
\alias{eye}
\alias{eye_in_speech_bubble}
\alias{brain}
\alias{tongue}
\alias{mouth}
\title{Body emojis}
\usage{
selfie()

flexed_biceps()

backhand_index_pointing_left()

backhand_index_pointing_right()

backhand_index_pointing_down()

backhand_index_pointing_up()

index_pointing_up()

middle_finger()

victory_hand()

crossed_fingers()

vulcan_salute()

sign_of_the_horns()

call_me_hand()

hand_with_fingers_splayed()

raised_hand()

thumbs_up()

thumbs_down()

raised_fist()

oncoming_fist()

left_facing_fist()

right_facing_fist()

raised_back_of_hand()

waving_hand()

love_you_gesture()

writing_hand()

clapping_hands()

open_hands()

raising_hands()

palms_up_together()

folded_hands()

handshake()

nail_polish()

ear()

nose()

footprints()

eyes()

eye()

eye_in_speech_bubble()

brain()

tongue()

mouth()
}
\description{
Body emojis
}
