% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{couple_with_heart}
\alias{couple_with_heart}
\title{Emoji sequence for couple with heart}
\usage{
couple_with_heart(x, y)
}
\arguments{
\item{x}{man or woman}

\item{y}{man or woman}
}
\description{
Emoji sequence for couple with heart
}
