% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{couple_with_heart}
\alias{couple_with_heart}
\alias{kiss}
\title{Family emoji sequences}
\usage{
couple_with_heart(x = woman, y = man)

kiss(x = man, y = woman)
}
\arguments{
\item{x}{man or woman}

\item{y}{man or woman}
}
\description{
Family emoji sequences
}
\examples{
\dontrun{
couple_with_heart( man, woman)
kiss( man, man )
}

}
