% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emojirx.R
\name{from_runes}
\alias{from_runes}
\title{convert space separated runes into their character representation}
\usage{
from_runes(runes)
}
\arguments{
\item{runes}{vector of space separated runes}
}
\description{
convert space separated runes into their character representation
}
\examples{
runes_to_char( "U+1F468" )

}
