% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji.R
\name{parse_emoji_list}
\alias{parse_emoji_list}
\title{Parse Emoji List file}
\usage{

  parse_emoji_list(emoji_list = "http://unicode.org/emoji/charts/emoji-list.html",
  full_emoji_list = "http://unicode.org/emoji/charts/full-emoji-list.html")
}
\arguments{
\item{url}{url for emoji list file}
}
\description{
Parse Emoji List file
}
