% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{runes_to_char}
\alias{runes_to_char}
\title{convert space separated runes into their character representation}
\usage{
runes_to_char(runes)
}
\arguments{
\item{runes}{vector of space separated runes}
}
\description{
convert space separated runes into their character representation
}
\examples{
runes_to_char( "U+1F468" )

}
