% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tones.R
\name{light}
\alias{light}
\alias{medium_light}
\alias{medium}
\alias{medium_dark}
\alias{dark}
\title{Skin tone modifier functions}
\usage{
light(txt)

medium_light(txt)

medium(txt)

medium_dark(txt)

dark(txt)
}
\arguments{
\item{txt}{An emoji to alter skin tone to}
}
\value{
an emoji with modified skin tone (if appropriate)
}
\description{
Skin tone modifier functions
}
