% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tones.R
\name{light_skin_tone}
\alias{light_skin_tone}
\alias{medium_light_skin_tone}
\alias{medium_skin_tone}
\alias{medium_dark_skin_tone}
\alias{dark_skin_tone}
\alias{skin_tone}
\title{Skin tone modifier functions}
\usage{
light_skin_tone(txt)

medium_light_skin_tone(txt)

medium_skin_tone(txt)

medium_dark_skin_tone(txt)

dark_skin_tone(txt)

skin_tone(txt, tone = c("light", "medium_light", "medium", "medium_dark",
  "dark"))
}
\arguments{
\item{txt}{An emoji to alter skin tone to}

\item{tone}{tone to apply}
}
\value{
an emoji with modified skin tone (if appropriate)
}
\description{
Skin tone modifier functions
}
